feat(easy-admin): add laySearch support for search boxes in lists
Introduce the `laySearch` option to enable the search functionality on list
page search boxes. This improvement allows users to have a more intuitive
and interactive way of searching through the list data by providing a search
box with clear and concise feedback.

When `laySearch` is set to true, the corresponding search box will now
be equipped with the `lay-search` class which enhances the user interface
and experience.